:truck: - :fire: - Deleting Next api in favour of own
- Running a server that parses these files into something a bit smaller
is not something that is part of Next's default suite. Instead of
figuring out how to roll their own server, I decided to run with what I
know, express. On startup, the server parses the input file to an array
/ tuple based type, bound by a contract set in `shared` that is loaded
into both the BE and FE via yarn workspaces. It's not insanely fast, but
can be made faster with workers at some point.
Fetching is fast though, instead of fetching 41megs of JSON, we now
fetch a mere 1.8mb (still a decent amount, but a lot less).